requirements.txt
# Core frameworks
torch==2.0.1
transformers==4.30.2
datasets==2.12.0
accelerate==0.20.3

# Reinforcement Learning
stable-baselines3==2.0.0
trl==0.4.7

# Data processing
pandas==2.0.2
numpy==1.24.3
beautifulsoup4==4.12.2
lxml==4.9.2

# Evaluation
scikit-learn==1.2.2
scipy==1.10.1

# Experiment management
wandb==0.15.3
tensorboard==2.13.0

# API service
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8

# Utilities
tqdm==4.65.0
pyyaml==6.0
python-dotenv==1.0.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Optional optimization
deepspeed==0.9.2

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
graphviz==0.20.1

# PEFT (LoRA)
peft==0.4.0
